general main window and applications page design
Designed the general appearance of the main window UI. This includes the navigation menu design, 3 contentless pages (only a page title), and 1 page (applications page) fully designed.

The applications page design comprises of a page title, table to display data from a database, and a filter to filter the table results.